Added fix for #56
